Wizard Walk
Added wizard walk animation testable in the scene "AnimationTest." The PlayerMove script now requires two additional variables for animation purposes: sprite renderer and animator. for now I'm using square for the sprite field in sprite renderer and am using the wizard animator controller found under assets/animations/wizard for the controller variable in animator. The current implementation flips the sprite based on movement commands so as not to duplicate animations based on the direction the player is facing.